Created many of the plugin schema files.
This was done in anticipation of making plugins installable from the /install url.  That install functionality will be for keeping zuha as light as possibleand only adding what is needed to be added.  It keeps the database lighter, especially the acos table, and makes the site perform noticeably faster.

Also, forcing a rename in this commit.